smol refactor and renaming
