+++ bin/reth/src/args/pruning_args.rs
@@ -2,7 +2,7 @@
 
 use clap::Args;
 use reth_config::config::PruneConfig;
-use reth_primitives::{ChainSpec, ContractLogPart, PruneMode, PruneModes};
+use reth_primitives::{ChainSpec, ContractLogsPruneConfig, PruneMode, PruneModes};
 use std::sync::Arc;
 
 /// Parameters for pruning and full node
@@ -31,7 +31,7 @@ impl PruningArgs {
                         .map(|contract| PruneMode::Before(contract.block)),
                     account_history: Some(PruneMode::Distance(128)),
                     storage_history: Some(PruneMode::Distance(128)),
-                    only_contract_logs: ContractLogPart(
+                    contract_logs_filter: ContractLogsPruneConfig(
                         chain_spec
                             .deposit_contract
                             .as_ref()
